// backend/mcp-salesforce-intelligence/src/shared/types.ts
export interface KBPayload {
  content: string;
  type: 'CONTEXT';
  tags: string;
  summary: string;
}